# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertReportEnvelope": {
        "properties": {
          "packet": {
            "$ref": "#/components/schemas/AlertReportPacket"
          },
          "report_id": {
            "title": "Report Id",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "packet"
        ],
        "title": "AlertReportEnvelope",
        "type": "object"
      },
      "AlertReportPacket": {
        "description": "Per-alert report — stored in ``incident_reports.summary`` JSONB.",
        "properties": {
          "alert_id": {
            "title": "Alert Id",
            "type": "integer"
          },
          "analyst": {
            "$ref": "#/components/schemas/AnalystSection"
          },
          "detection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectionSection"
              },
              {
                "type": "null"
              }
            ]
          },
          "disposition": {
            "$ref": "#/components/schemas/AlertDisposition"
          },
          "final_summary": {
            "title": "Final Summary",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "investigation": {
            "$ref": "#/components/schemas/InvestigationSection"
          },
          "kind": {
            "$ref": "#/components/schemas/IncidentKind",
            "default": "PER_ALERT"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "overview": {
            "$ref": "#/components/schemas/OverviewSection"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "response": {
            "$ref": "#/components/schemas/ResponseSection"
          },
          "severity_priority": {
            "$ref": "#/components/schemas/SeverityPrioritySection"
          },
          "timeline": {
            "$ref": "#/components/schemas/TimelineSection"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workflow_status": {
            "$ref": "#/components/schemas/AlertStatus"
          }
        },
        "required": [
          "alert_id",
          "title",
          "generated_at",
          "workflow_status",
          "disposition",
          "overview",
          "severity_priority",
          "investigation",
          "timeline",
          "response",
          "analyst",
          "final_summary",
          "markdown"
        ],
        "title": "AlertReportPacket",
        "type": "object"
      },
...
      "AnalystSection": {
        "properties": {
          "disposition": {
            "$ref": "#/components/schemas/AlertDisposition"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AnalystEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          }
        },
        "required": [
          "status",
          "disposition"
        ],
        "title": "AnalystSection",
        "type": "object"
      },
...
      "DailySummaryEnvelope": {
        "properties": {
          "packet": {
            "$ref": "#/components/schemas/DailySummaryPacket"
          },
          "report_id": {
            "title": "Report Id",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "packet"
        ],
        "title": "DailySummaryEnvelope",
        "type": "object"
      },
      "DailySummaryPacket": {
        "description": "Daily roll-up — also stored in ``incident_reports.summary`` JSONB.",
        "properties": {
          "by_disposition": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Disposition",
            "type": "object"
          },
          "by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Severity",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Status",
            "type": "object"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "final_summary": {
            "title": "Final Summary",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/IncidentKind",
            "default": "DAILY_SUMMARY"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "mean_investigation_latency_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Investigation Latency Seconds"
          },
          "mean_report_latency_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Report Latency Seconds"
          },
          "mean_response_latency_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Response Latency Seconds"
          },
          "mean_triage_latency_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Triage Latency Seconds"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "response_actions_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Response Actions By Status",
            "type": "object"
          },
          "response_actions_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Response Actions By Type",
            "type": "object"
          },
          "response_actions_total": {
            "default": 0,
            "title": "Response Actions Total",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "top_destinations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Destinations",
            "type": "array"
          },
          "top_predictions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Predictions",
            "type": "array"
          },
          "top_sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Top Sources",
            "type": "array"
          },
          "total_alerts": {
            "title": "Total Alerts",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "generated_at",
          "date",
          "period_start",
          "period_end",
          "total_alerts",
          "final_summary",
          "markdown"
        ],
        "title": "DailySummaryPacket",
        "type": "object"
      },
      "DailySummaryRequest": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          }
        },
        "title": "DailySummaryRequest",
        "type": "object"
      },
...
      "DetectionSection": {
        "properties": {
          "class_probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Class Probabilities",
            "type": "object"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "predicted_label": {
            "title": "Predicted Label",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "required": [
          "predicted_label",
          "confidence"
        ],
        "title": "DetectionSection",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentKind": {
        "enum": [
          "PER_ALERT",
          "DAILY_SUMMARY"
        ],
        "title": "IncidentKind",
        "type": "string"
      },
      "IncidentReportOut": {
        "description": "Lightweight summary row for ``GET /api/v1/reports``.",
        "properties": {
          "alert_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/IncidentKind"
          },
          "md_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Md Path"
          },
          "pdf_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Path"
          },
          "period_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "alert_id",
          "title",
          "period_start",
          "period_end",
          "md_path",
          "pdf_path",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentReportOut",
        "type": "object"
      },
...
      "ReportRequest": {
        "description": "Body for ``POST /api/v1/alerts/{id}/report`` — currently no knobs.",
        "properties": {},
        "title": "ReportRequest",
        "type": "object"
      },
...
      "TaskKind": {
        "description": "Kind of background job tracked in the ``tasks`` table.",
        "enum": [
          "DETECTION_RUN",
          "REPORT_ALERT",
          "DAILY_SUMMARY",
          "DRIFT_RUN",
          "RETENTION_CLEANUP",
          "ML_RETRAIN"
        ],
        "title": "TaskKind",
        "type": "string"
      },
...
      "TaskOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TaskKind"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "progress",
          "params",
          "result",
          "error",
          "created_by",
          "created_at",
          "updated_at",
          "started_at",
          "finished_at"
        ],
        "title": "TaskOut",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCEEDED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI-driven intrusion detection and response dashboard.",
    "title": "SentinelAI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/alerts/{alert_id}/report": {
      "get": {
        "description": "Return the most recent incident report for this alert (no re-generation).",
        "operationId": "get_alert_report_endpoint_api_v1_alerts__alert_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertReportEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Alert Report Endpoint",
        "tags": [
          "alerts"
        ]
      },
      "post": {
        "description": "Generate (and persist) an incident report for this alert.",
        "operationId": "generate_alert_report_endpoint_api_v1_alerts__alert_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReportRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertReportEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Alert Report Endpoint",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/api/v1/reports": {
      "get": {
        "operationId": "list_reports_endpoint_api_v1_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IncidentKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "alert_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alert Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentReportOut"
                  },
                  "title": "Response List Reports Endpoint Api V1 Reports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reports Endpoint",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/daily/run": {
      "post": {
        "operationId": "daily_run_api_v1_reports_daily_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DailySummaryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySummaryEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily Run",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "operationId": "get_report_endpoint_api_v1_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Report Endpoint Api V1 Reports  Report Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Endpoint",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{report_id}/markdown": {
      "get": {
        "operationId": "get_report_markdown_api_v1_reports__report_id__markdown_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Markdown",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/v1/tasks/report/{alert_id}": {
      "post": {
        "operationId": "enqueue_alert_report_api_v1_tasks_report__alert_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "sentinelai_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentinelai Access"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enqueue Alert Report",
        "tags": [
          "tasks"
        ]
      }
    },
...
  }
}